+++ vector/src/main/java/im/vector/riotx/features/invite/InviteUsersToRoomActivity.kt
@@ -36,8 +36,8 @@ import im.vector.riotx.core.platform.WaitingViewData
 import im.vector.riotx.core.utils.toast
 import im.vector.riotx.features.userdirectory.KnownUsersFragment
 import im.vector.riotx.features.userdirectory.KnownUsersFragmentArgs
-import im.vector.riotx.features.userdirectory.PhoneBookFragment
-import im.vector.riotx.features.userdirectory.PhoneBookViewModel
+import im.vector.riotx.features.phonebook.PhoneBookFragment
+import im.vector.riotx.features.phonebook.PhoneBookViewModel
 import im.vector.riotx.features.userdirectory.UserDirectoryFragment
 import im.vector.riotx.features.userdirectory.UserDirectorySharedAction
 import im.vector.riotx.features.userdirectory.UserDirectorySharedActionViewModel
